refactor(#3038): extract health.rs send-to-agent dispatch to outbound/send_to_agent.rs
Pure, behavior-preserving move of the agent-to-agent relay entry point out of
the health module, which had fused send-to-agent dispatch with health-check
logic (EPIC #3038 god-object slice).

Extracted cohesive cluster (from src/services/discord/health.rs):
- ParsedSendToAgentRequest (struct)
- parse_send_to_agent_body (request parser)
- handle_send_to_agent (agent:<role_id> router)

into new module src/services/discord/outbound/send_to_agent.rs. Delivery still
flows through the unchanged pub(crate) health::send_message_with_backends; no
new global state, queue, or lease. Call sites re-pointed with identical args:
- src/services/auto_queue/route_request_generate.rs
- src/server/routes/health_api.rs

health.rs: 2369 -> 2293 production LoC (giant baseline ratcheted down,
change-surfaces freeze + ARCHITECTURE + multinode Audited touches synced).
Added 7 unit tests for the parser (no test previously covered it).

Gates: cargo fmt --check clean, cargo check --lib / --lib --tests clean (0 new
warnings), cargo test --lib (send_to_agent 7 + health 42 pass),
generate_inventory_docs + check_agent_maintenance_docs freshness passed,
ci-script-checks exit 0.

Refs #3038
