+++ ARCHITECTURE.md
@@ -398,6 +398,7 @@ src/
 │   │   │   ├── mod.rs
 │   │   │   ├── policy.rs
 │   │   │   ├── result.rs
+│   │   │   ├── send_to_agent.rs
 │   │   │   └── transport.rs
 │   │   ├── placeholder_live_events/
 │   │   │   ├── common.rs
