+++ src/services/discord/health.rs
@@ -2203,82 +2203,6 @@ pub async fn handle_send<'a>(
     .await
 }
 
-#[derive(Debug, PartialEq, Eq)]
-struct ParsedSendToAgentRequest {
-    role_id: String,
-    message: String,
-    mode: String,
-}
-
-#[cfg_attr(not(test), allow(dead_code))]
-fn parse_send_to_agent_body(body: &str) -> Result<ParsedSendToAgentRequest, &'static str> {
-    let json: serde_json::Value = serde_json::from_str(body).map_err(|_| "invalid JSON")?;
-    let role_id = json
-        .get("role_id")
-        .and_then(|value| value.as_str())
-        .map(str::trim)
-        .unwrap_or("")
-        .to_string();
-    if role_id.is_empty() {
-        return Err("role_id is required");
-    }
-
-    let message = json
-        .get("message")
-        .and_then(|value| value.as_str())
-        .unwrap_or("")
-        .to_string();
-    if message.is_empty() {
-        return Err("message is required");
-    }
-
-    let mode = json
-        .get("mode")
-        .and_then(|value| value.as_str())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .unwrap_or("announce");
-    if !matches!(mode, "announce" | "notify") {
-        return Err("mode must be announce or notify");
-    }
-
-    Ok(ParsedSendToAgentRequest {
-        role_id,
-        message,
-        mode: mode.to_string(),
-    })
-}
-
-pub async fn handle_send_to_agent(
-    registry: &HealthRegistry,
-    sqlite: Option<&Db>,
-    pg_pool: Option<&PgPool>,
-    body: &str,
-) -> (&'static str, String) {
-    let request = match parse_send_to_agent_body(body) {
-        Ok(request) => request,
-        Err(error) => {
-            return (
-                "400 Bad Request",
-                serde_json::json!({"ok": false, "error": error}).to_string(),
-            );
-        }
-    };
-
-    let target = format!("agent:{}", request.role_id);
-    send_message_with_backends(
-        registry,
-        sqlite,
-        pg_pool,
-        &target,
-        &request.message,
-        "system",
-        &request.mode,
-        None,
-    )
-    .await
-}
-
 /// Handle POST /api/discord/send-dm — send a DM to a Discord user.
 /// Accepts JSON:
 /// {"user_id":"...", "content":"...", "bot":"announce|notify|claude|codex"}
